refactor: extract MultipleChoiceQuestion to separate component
- Create new MultipleChoiceQuestion.tsx component file
- Update QuizSection to import and use the extracted component
- Improves code organization and modularity
