+++ apps/nextjs/src/components/AppComponents/SectionContent/QuizSection/index.tsx
@@ -1,86 +1,32 @@
-import type {
-  QuizV1,
-  QuizV2QuestionMultipleChoice,
-} from "@oakai/aila/src/protocol/schema";
+import { useMemo } from "react";
+
+import type { LessonPlanSectionWhileStreaming } from "@oakai/aila/src/protocol/schema";
+import { QuizV1Schema } from "@oakai/aila/src/protocol/schema";
 import { convertQuizV1ToV2 } from "@oakai/aila/src/protocol/schemas/quiz/conversion/quizV1ToV2";
 
-import { OakBox, OakFlex, OakIcon, OakP } from "@oaknational/oak-components";
-
-import { shuffleMultipleChoiceAnswers } from "./shuffle";
+import { MultipleChoiceQuestion } from "./MultipleChoiceQuestion";
 
 export type QuizSectionProps = {
-  quiz: QuizV1;
+  // When we have agentic generation, we will know that sections are valid when streamed
+  // Until then, it's a loose type
+  quizSection: LessonPlanSectionWhileStreaming;
 };
 
-type MultipleChoiceQuestionProps = {
-  question: QuizV2QuestionMultipleChoice;
-};
+export const QuizSection = ({ quizSection }: QuizSectionProps) => {
+  const quiz = useMemo(() => {
+    return QuizV1Schema.safeParse(quizSection).data;
+  }, [quizSection]);
 
-const MultipleChoiceQuestion = ({
-  question,
-  questionNumber,
-}: MultipleChoiceQuestionProps & { questionNumber: number }) => {
-  const answers = shuffleMultipleChoiceAnswers(
-    question.answers,
-    question.distractors,
-  );
+  if (!quiz) {
+    // Shouldn't happen, but just to be safe
+    return "Invalid quiz";
+  }
 
-  const correctAnswerCount = question.answers.length;
-
-  const ensureEndsWithPeriod = (text: string) => {
-    const endsWithPunctuation = /[.!?]$/.test(text);
-    return endsWithPunctuation ? text : `${text}.`;
-  };
-
-  const tickInstruction = `Tick ${correctAnswerCount} correct answer${correctAnswerCount !== 1 ? "s" : ""}.`;
-  const fullQuestionText = `${questionNumber}. ${ensureEndsWithPeriod(question.question)} ${tickInstruction}`;
-
-  return (
-    <OakBox $mb="space-between-l">
-      <OakP $mb="space-between-s">{fullQuestionText}</OakP>
-
-      {question.hint && (
-        <OakP $mb="space-between-s" $color="text-subdued">
-          {question.hint}
-        </OakP>
-      )}
-
-      <OakBox>
-        {answers.map((answer, index) => (
-          <OakFlex key={index} $alignItems="center" $mb="space-between-xs">
-            <OakBox
-              $mr="space-between-xs"
-              $width="all-spacing-7"
-              $height="all-spacing-7"
-              $ba="border-solid-m"
-              $borderColor="black"
-            >
-              {answer.isCorrect && (
-                <OakIcon
-                  iconName="tick"
-                  $width="100%"
-                  $height="100%"
-                  $color="text-inverted"
-                  $transform="scale(1.15)"
-                />
-              )}
-            </OakBox>
-            <OakP $font={answer.isCorrect ? "body-2-bold" : "body-2"}>
-              {String.fromCharCode(97 + index)}) {answer.text}
-            </OakP>
-          </OakFlex>
-        ))}
-      </OakBox>
-    </OakBox>
-  );
-};
-
-export const QuizSection = ({ quiz }: QuizSectionProps) => {
   const quizV2 = convertQuizV1ToV2(quiz);
-  const questions = quizV2.questions;
+
   return (
     <>
-      {questions.map((question, index) => {
+      {quizV2.questions.map((question, index) => {
         if (question.questionType === "multiple-choice") {
           return (
             <MultipleChoiceQuestion
